Refactor Validation classes
