+++ src/main/java/app/MoneyGoWhere.java
@@ -2,11 +2,10 @@ package app;
 
 import item.Item;
 import item.ItemList;
-import order.OrderEntry;
-import validation.ItemValidation;
+import validation.item.AddItemValidation;
+import validation.item.DeleteItemValidation;
 import order.Order;
 import order.OrderList;
-import validation.ItemValidation;
 import utility.Parser;
 import utility.Ui;
 import exception.InvalidArgumentException;
@@ -29,21 +28,22 @@ public class MoneyGoWhere {
 
     public void handleCommand(Command command) throws InvalidArgumentException {
         Ui ui = new Ui();
-        ItemValidation itemValidation = new ItemValidation();
+        AddItemValidation addItemValidation = new AddItemValidation();
+        DeleteItemValidation deleteItemValidation = new DeleteItemValidation();
         switch (command.getCommand()) {
         case "listitem":
             items.displayList();
             break;
         case "additem":
             //Print some header
-            if (!itemValidation.isValid(command)) {
+            if (!addItemValidation.isValidFormat(command)) {
                 break;
             }
 
             command.mapArgumentAlias("name", "n");
             command.mapArgumentAlias("price", "p");
 
-            if (!itemValidation.isValid(command)) {
+            if (!addItemValidation.isValid(command)) {
                 break;
             }
 
@@ -60,13 +60,13 @@ public class MoneyGoWhere {
         case "deleteitem":
             command.mapArgumentAlias("index", "i");
 
-            if (!itemValidation.isValidFormatDelete(command)) {
+            if (!deleteItemValidation.isValidFormat(command)) {
                 break;
             }
-            if (!itemValidation.isInteger(command.getArgumentMap().get("index"))) {
+            if (!deleteItemValidation.isInteger(command.getArgumentMap().get("index"))) {
                 break;
             }
-            if (!itemValidation.isValidIndex(command.getArgumentMap().get("index"), items)) {
+            if (!deleteItemValidation.isValidIndex(command.getArgumentMap().get("index"), items)) {
                 break;
             }
